clean up QuantumCell
