+++ halo2-base/src/gates/flex_gate.rs
@@ -1,7 +1,3 @@
-use super::{
-    AssignedValue, Context, GateInstructions,
-    QuantumCell::{self, Constant, Existing, Witness},
-};
 use crate::halo2_proofs::{
     circuit::Value,
     plonk::{
@@ -11,6 +7,10 @@ use crate::halo2_proofs::{
     poly::Rotation,
 };
 use crate::utils::ScalarField;
+use crate::{
+    AssignedValue, Context,
+    QuantumCell::{self, Constant, Existing, Witness, WitnessFraction},
+};
 use itertools::Itertools;
 use std::{
     iter::{self, once},
@@ -185,8 +185,8 @@ impl<F: ScalarField> FlexGateConfig<F> {
     pub fn inner_product_simple<'a>(
         &self,
         ctx: &mut Context<F>,
-        a: impl IntoIterator<Item = QuantumCell<'a, F>>,
-        b: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
     ) -> AssignedValue<F> {
         let mut sum;
         let mut a = a.into_iter();
@@ -216,8 +216,8 @@ impl<F: ScalarField> FlexGateConfig<F> {
     pub fn inner_product_simple_with_assignments<'a>(
         &self,
         ctx: &mut Context<F>,
-        a: impl IntoIterator<Item = QuantumCell<'a, F>>,
-        b: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
     ) -> (Vec<AssignedValue<F>>, AssignedValue<F>) {
         let mut sum;
         let mut a = a.into_iter();
@@ -249,8 +249,8 @@ impl<F: ScalarField> FlexGateConfig<F> {
     fn inner_product_with_assignments<'a, 'b: 'a>(
         &self,
         ctx: &mut Context<F>,
-        a: impl IntoIterator<Item = QuantumCell<'a, F>>,
-        b: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
     ) -> (Vec<AssignedValue<F>>, AssignedValue<F>) {
         // we will do special handling of the cases where one of the vectors is all constants
         match self.strategy {
@@ -320,6 +320,655 @@ impl<F: ScalarField> FlexGateConfig<F> {
     }
 }
 
+pub trait GateInstructions<F: ScalarField> {
+    fn strategy(&self) -> GateStrategy;
+    fn context_id(&self) -> usize;
+
+    fn pow_of_two(&self) -> &[F];
+    fn get_field_element(&self, n: u64) -> F;
+
+    fn assign_region<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
+        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
+    ) -> Vec<AssignedValue<F>> {
+        self.assign_region_in(ctx, inputs, gate_offsets, ctx.current_phase())
+    }
+
+    fn assign_region_in<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
+        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
+        phase: usize,
+    ) -> Vec<AssignedValue<F>>;
+
+    /// Only returns the last assigned cell
+    ///
+    /// Does not collect the vec, saving heap allocation
+    fn assign_region_last<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
+        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
+    ) -> AssignedValue<F> {
+        self.assign_region_last_in(ctx, inputs, gate_offsets, ctx.current_phase())
+    }
+
+    fn assign_region_last_in<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
+        gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
+        phase: usize,
+    ) -> AssignedValue<F>;
+
+    /// Only call this if ctx.region is not in shape mode, i.e., if not using simple layouter or ctx.first_pass = false
+    ///
+    /// All indices in `gate_offsets`, `equality_offsets`, `external_equality` are with respect to `inputs` indices
+    /// - `gate_offsets` specifies indices to enable selector for the gate; assume `gate_offsets` is sorted in increasing order
+    /// - `equality_offsets` specifies pairs of indices to constrain equality
+    /// - `external_equality` specifies an existing cell to constrain equality with the cell at a certain index
+    fn assign_region_smart<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
+        gate_offsets: impl IntoIterator<Item = usize>,
+        equality_offsets: impl IntoIterator<Item = (usize, usize)>,
+        external_equality: Vec<(&AssignedValue<F>, usize)>,
+    ) -> Vec<AssignedValue<F>> {
+        let assignments =
+            self.assign_region(ctx, inputs, gate_offsets.into_iter().map(|i| (i as isize, None)));
+        for (offset1, offset2) in equality_offsets.into_iter() {
+            ctx.region
+                .constrain_equal(assignments[offset1].cell(), assignments[offset2].cell())
+                .unwrap();
+        }
+        for (assigned, eq_offset) in external_equality.into_iter() {
+            ctx.region.constrain_equal(assigned.cell(), assignments[eq_offset].cell()).unwrap();
+        }
+        assignments
+    }
+
+    fn assign_witnesses(
+        &self,
+        ctx: &mut Context<F>,
+        witnesses: impl IntoIterator<Item = Value<F>>,
+    ) -> Vec<AssignedValue<F>> {
+        self.assign_region(ctx, witnesses.into_iter().map(Witness), [])
+    }
+
+    fn load_witness(&self, ctx: &mut Context<F>, witness: Value<F>) -> AssignedValue<F> {
+        self.assign_region_last(ctx, [Witness(witness)], [])
+    }
+
+    fn load_constant(&self, ctx: &mut Context<F>, c: F) -> AssignedValue<F> {
+        self.assign_region_last(ctx, [Constant(c)], [])
+    }
+
+    fn load_zero(&self, ctx: &mut Context<F>) -> AssignedValue<F> {
+        if let Some(zcell) = &ctx.zero_cell {
+            return zcell.clone();
+        }
+        let zero_cell = self.assign_region_last(ctx, [Constant(F::zero())], []);
+        ctx.zero_cell = Some(zero_cell.clone());
+        zero_cell
+    }
+
+    /// Copies a, b and constrains `a + b * 1 = out`
+    // | a | b | 1 | a + b |
+    fn add(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+        let out_val = a.value().zip(b.value()).map(|(a, b)| *a + b);
+        self.assign_region_last(
+            ctx,
+            vec![a, b, Constant(F::one()), Witness(out_val)],
+            vec![(0, None)],
+        )
+    }
+
+    /// Copies a, b and constrains `a + b * (-1) = out`
+    // | a - b | b | 1 | a |
+    fn sub(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+
+        let out_val = a.value().zip(b.value()).map(|(a, b)| *a - b);
+        // slightly better to not have to compute -F::one() since F::one() is cached
+        let assigned_cells = self.assign_region(
+            ctx,
+            vec![Witness(out_val), b, Constant(F::one()), a],
+            vec![(0, None)],
+        );
+        assigned_cells.into_iter().next().unwrap()
+    }
+
+    // | a | -a | 1 | 0 |
+    fn neg(&self, ctx: &mut Context<F>, a: impl Into<QuantumCell<F>>) -> AssignedValue<F> {
+        let a = a.into();
+        let out_val = a.value().map(|v| -*v);
+        let assigned_cells = self.assign_region(
+            ctx,
+            vec![a, Witness(out_val), Constant(F::one()), Constant(F::zero())],
+            vec![(0, None)],
+        );
+        assigned_cells.into_iter().nth(1).unwrap()
+    }
+
+    /// Copies a, b and constrains `0 + a * b = out`
+    // | 0 | a | b | a * b |
+    fn mul(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+        let out_val = a.value().zip(b.value()).map(|(a, b)| *a * b);
+        self.assign_region_last(
+            ctx,
+            vec![Constant(F::zero()), a, b, Witness(out_val)],
+            vec![(0, None)],
+        )
+    }
+
+    /// a * b + c
+    fn mul_add(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+        c: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+        let c = c.into();
+        let out_val = a.value().zip(b.value()).map(|(a, b)| *a * b) + c.value();
+        self.assign_region_last(ctx, vec![c, a, b, Witness(out_val)], vec![(0, None)])
+    }
+
+    /// (1 - a) * b = b - a * b
+    fn mul_not(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+
+        let out_val = a.value().zip(b.value()).map(|(a, b)| (F::one() - a) * b);
+        let assignments =
+            self.assign_region(ctx, vec![Witness(out_val), a, b.clone(), b], vec![(0, None)]);
+        ctx.region.constrain_equal(assignments[2].cell(), assignments[3].cell()).unwrap();
+        assignments.into_iter().next().unwrap()
+    }
+
+    /// Constrain x is 0 or 1.
+    fn assert_bit(&self, ctx: &mut Context<F>, x: AssignedValue<F>) {
+        self.assign_region_last(
+            ctx,
+            [Constant(F::zero()), Existing(x), Existing(x), Existing(x)],
+            [(0, None)],
+        );
+    }
+
+    fn div_unsafe(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+
+        // TODO: if really necessary, make `c` of type `Assigned<F>`
+        // this would require the API using `Assigned<F>` instead of `F` everywhere, so leave as last resort
+        let c = a.value().zip(b.value()).map(|(a, b)| b.invert().unwrap() * a);
+        let assignments =
+            self.assign_region(ctx, vec![Constant(F::zero()), Witness(c), b, a], vec![(0, None)]);
+        assignments.into_iter().nth(1).unwrap()
+    }
+
+    fn assert_equal(&self, ctx: &mut Context<F>, a: QuantumCell<F>, b: QuantumCell<F>) {
+        if let (Existing(a), Existing(b)) = (&a, &b) {
+            ctx.region.constrain_equal(a.cell(), b.cell()).unwrap();
+        } else {
+            self.assign_region_smart(
+                ctx,
+                vec![Constant(F::zero()), a, Constant(F::one()), b],
+                vec![0],
+                vec![],
+                vec![],
+            );
+        }
+    }
+
+    fn assert_is_const(&self, ctx: &mut Context<F>, a: &AssignedValue<F>, constant: F) {
+        let c_cell = ctx.assign_fixed(constant);
+        #[cfg(feature = "halo2-axiom")]
+        ctx.region.constrain_equal(a.cell(), &c_cell);
+        #[cfg(feature = "halo2-pse")]
+        ctx.region.constrain_equal(a.cell(), c_cell).unwrap();
+    }
+
+    /// Returns `(assignments, output)` where `output` is the inner product of `<a, b>`
+    ///
+    /// `assignments` is for internal use
+    fn inner_product<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
+    ) -> AssignedValue<F>;
+
+    /// very specialized for optimal range check, not for general consumption
+    /// - `a_assigned` is expected to have capacity a.len()
+    /// - we re-use `a_assigned` to save memory allocation
+    fn inner_product_left<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
+        a_assigned: &mut Vec<AssignedValue<F>>,
+    ) -> AssignedValue<F>;
+
+    /// Returns an iterator with the partial sums `sum_{j=0..=i} a[j] * b[j]`.
+    fn inner_product_with_sums<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
+    ) -> Box<dyn Iterator<Item = AssignedValue<F>>>;
+
+    fn sum<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let mut a = a.into_iter().peekable();
+        let start = a.next();
+        if start.is_none() {
+            return self.load_zero(ctx);
+        }
+        let start = start.unwrap();
+        if a.peek().is_none() {
+            return self.assign_region_last(ctx, [start], []);
+        }
+        let (len, hi) = a.size_hint();
+        debug_assert_eq!(Some(len), hi);
+
+        let mut sum = start.value().copied();
+        let cells = iter::once(start).chain(a.flat_map(|a| {
+            sum = sum + a.value();
+            [a, Constant(F::one()), Witness(sum)]
+        }));
+        self.assign_region_last(ctx, cells, (0..len).map(|i| (3 * i as isize, None)))
+    }
+
+    /// Returns the assignment trace where `output[3 * i]` has the running sum `sum_{j=0..=i} a[j]`
+    fn sum_with_assignments<'a>(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+    ) -> Vec<AssignedValue<F>> {
+        let mut a = a.into_iter().peekable();
+        let start = a.next();
+        if start.is_none() {
+            return vec![self.load_zero(ctx)];
+        }
+        let start = start.unwrap();
+        if a.peek().is_none() {
+            return self.assign_region(ctx, [start], []);
+        }
+        let (len, hi) = a.size_hint();
+        debug_assert_eq!(Some(len), hi);
+
+        let mut sum = start.value().copied();
+        let cells = iter::once(start).chain(a.flat_map(|a| {
+            sum = sum + a.value();
+            [a, Constant(F::one()), Witness(sum)]
+        }));
+        self.assign_region(ctx, cells, (0..len).map(|i| (3 * i as isize, None)))
+    }
+
+    // requires b.len() == a.len() + 1
+    // returns
+    // x_i = b_1 * (a_1...a_{i - 1})
+    //     + b_2 * (a_2...a_{i - 1})
+    //     + ...
+    //     + b_i
+    // Returns [x_1, ..., x_{b.len()}]
+    fn accumulated_product(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
+    ) -> Vec<AssignedValue<F>> {
+        let mut b = b.into_iter();
+        let mut a = a.into_iter();
+        let b_first = b.next();
+        if let Some(b_first) = b_first {
+            let b_first = self.assign_region_last(ctx, [b_first], []);
+            std::iter::successors(Some(b_first), |&x| {
+                a.next().zip(b.next()).map(|(a, b)| self.mul_add(ctx, Existing(x), a, b))
+            })
+            .collect()
+        } else {
+            vec![]
+        }
+    }
+
+    fn sum_products_with_coeff_and_var<'a, 'b: 'a>(
+        &self,
+        ctx: &mut Context<F>,
+        values: impl IntoIterator<Item = (F, QuantumCell<F>, QuantumCell<F>)>,
+        var: QuantumCell<F>,
+    ) -> AssignedValue<F>;
+
+    // | 1 - b | 1 | b | 1 | b | a | 1 - b | out |
+    fn or(&self, ctx: &mut Context<F>, a: QuantumCell<F>, b: QuantumCell<F>) -> AssignedValue<F> {
+        let not_b_val = b.value().map(|x| F::one() - x);
+        let out_val = a.value().zip(b.value()).map(|(a, b)| *a + b)
+            - a.value().zip(b.value()).map(|(a, b)| *a * b);
+        let cells = vec![
+            Witness(not_b_val),
+            Constant(F::one()),
+            b.clone(),
+            Constant(F::one()),
+            b,
+            a,
+            Witness(not_b_val),
+            Witness(out_val),
+        ];
+        let mut assigned_cells =
+            self.assign_region_smart(ctx, cells, vec![0, 4], vec![(0, 6), (2, 4)], vec![]);
+        assigned_cells.pop().unwrap()
+    }
+
+    // | 0 | a | b | out |
+    fn and(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        self.mul(ctx, a, b)
+    }
+
+    fn not(&self, ctx: &mut Context<F>, a: impl Into<QuantumCell<F>>) -> AssignedValue<F> {
+        self.sub(ctx, Constant(F::one()), a)
+    }
+
+    fn select(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+        sel: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F>;
+
+    fn or_and(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+        c: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F>;
+
+    /// assume bits has boolean values
+    /// returns vec[idx] with vec[idx] = 1 if and only if bits == idx as a binary number
+    fn bits_to_indicator(
+        &self,
+        ctx: &mut Context<F>,
+        bits: &[AssignedValue<F>],
+    ) -> Vec<AssignedValue<F>> {
+        let k = bits.len();
+
+        let (inv_last_bit, last_bit) = {
+            let mut assignments = self
+                .assign_region(
+                    ctx,
+                    vec![
+                        Witness(bits[k - 1].value().map(|b| F::one() - b)),
+                        Existing(bits[k - 1]),
+                        Constant(F::one()),
+                        Constant(F::one()),
+                    ],
+                    vec![(0, None)],
+                )
+                .into_iter();
+            (assignments.next().unwrap(), assignments.next().unwrap())
+        };
+        let mut indicator = Vec::with_capacity(2 * (1 << k) - 2);
+        let mut offset = 0;
+        indicator.push(inv_last_bit);
+        indicator.push(last_bit);
+        for (idx, bit) in bits.iter().rev().enumerate().skip(1) {
+            for old_idx in 0..(1 << idx) {
+                let inv_prod_val = indicator[offset + old_idx]
+                    .value()
+                    .zip(bit.value())
+                    .map(|(a, b)| (F::one() - b) * a);
+                let inv_prod = self
+                    .assign_region_smart(
+                        ctx,
+                        vec![
+                            Witness(inv_prod_val),
+                            Existing(indicator[offset + old_idx]),
+                            Existing(*bit),
+                            Existing(indicator[offset + old_idx]),
+                        ],
+                        vec![0],
+                        vec![],
+                        vec![],
+                    )
+                    .into_iter()
+                    .next()
+                    .unwrap();
+                indicator.push(inv_prod);
+
+                let prod = self.mul(ctx, Existing(indicator[offset + old_idx]), Existing(*bit));
+                indicator.push(prod);
+            }
+            offset += 1 << idx;
+        }
+        indicator.split_off((1 << k) - 2)
+    }
+
+    // returns vec with vec.len() == len such that:
+    //     vec[i] == 1{i == idx}
+    fn idx_to_indicator(
+        &self,
+        ctx: &mut Context<F>,
+        idx: impl Into<QuantumCell<F>>,
+        len: usize,
+    ) -> Vec<AssignedValue<F>> {
+        let mut idx = idx.into();
+        let ind = self.assign_region(
+            ctx,
+            (0..len).map(|i| {
+                Witness(idx.value().map(|x| {
+                    if x.get_lower_32() == i as u32 {
+                        F::one()
+                    } else {
+                        F::zero()
+                    }
+                }))
+            }),
+            vec![],
+        );
+
+        // check ind[i] * (i - idx) == 0
+        for (i, ind) in ind.iter().enumerate() {
+            let val = ind.value().zip(idx.value()).map(|(ind, idx)| *ind * idx);
+            let assignments = self.assign_region(
+                ctx,
+                vec![
+                    Constant(F::zero()),
+                    Existing(*ind),
+                    idx,
+                    Witness(val),
+                    Constant(-F::from(i as u64)),
+                    Existing(*ind),
+                    Constant(F::zero()),
+                ],
+                vec![(0, None), (3, None)],
+            );
+            // need to use assigned idx after i > 0 so equality constraint holds
+            idx = Existing(assignments.into_iter().nth(2).unwrap());
+        }
+        ind
+    }
+
+    // performs inner product on a, indicator
+    // `indicator` values are all boolean
+    /// Assumes for witness generation that only one element of `indicator` has non-zero value and that value is `F::one()`.
+    fn select_by_indicator(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        indicator: impl IntoIterator<Item = AssignedValue<F>>,
+    ) -> AssignedValue<F> {
+        let mut sum = Value::known(F::zero());
+        let a = a.into_iter();
+        let (len, hi) = a.size_hint();
+        debug_assert_eq!(Some(len), hi);
+
+        let cells =
+            std::iter::once(Constant(F::zero())).chain(a.zip(indicator).flat_map(|(a, ind)| {
+                sum = sum.zip(a.value().zip(ind.value())).map(|(sum, (a, ind))| {
+                    if ind.is_zero_vartime() {
+                        sum
+                    } else {
+                        *a
+                    }
+                });
+                [a, Existing(ind), Witness(sum)]
+            }));
+        self.assign_region_last(ctx, cells, (0..len).map(|i| (3 * i as isize, None)))
+    }
+
+    fn select_from_idx<'a, 'v: 'a>(
+        &self,
+        ctx: &mut Context<F>,
+        cells: impl IntoIterator<Item = QuantumCell<F>>,
+        idx: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let cells = cells.into_iter();
+        let (len, hi) = cells.size_hint();
+        debug_assert_eq!(Some(len), hi);
+
+        let ind = self.idx_to_indicator(ctx, idx, len);
+        let out = self.select_by_indicator(ctx, cells, ind);
+        out
+    }
+
+    // | out | a | inv | 1 | 0 | a | out | 0
+    fn is_zero(&self, ctx: &mut Context<F>, a: &AssignedValue<F>) -> AssignedValue<F> {
+        let (is_zero, inv) = a
+            .value()
+            .map(|x| {
+                if x.is_zero_vartime() {
+                    (F::one(), Assigned::Trivial(F::one()))
+                } else {
+                    (F::zero(), Assigned::Rational(F::one(), *x))
+                }
+            })
+            .unzip();
+
+        let cells = vec![
+            Witness(is_zero),
+            Existing(*a),
+            WitnessFraction(inv),
+            Constant(F::one()),
+            Constant(F::zero()),
+            Existing(*a),
+            Witness(is_zero),
+            Constant(F::zero()),
+        ];
+        let assigned_cells = self.assign_region_smart(ctx, cells, vec![0, 4], vec![(0, 6)], vec![]);
+        assigned_cells.into_iter().next().unwrap()
+    }
+
+    fn is_equal(
+        &self,
+        ctx: &mut Context<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+    ) -> AssignedValue<F> {
+        let diff = self.sub(ctx, a, b);
+        self.is_zero(ctx, &diff)
+    }
+
+    // returns little-endian bit vectors
+    fn num_to_bits(
+        &self,
+        ctx: &mut Context<F>,
+        a: &AssignedValue<F>,
+        range_bits: usize,
+    ) -> Vec<AssignedValue<F>>;
+
+    /// given pairs `coords[i] = (x_i, y_i)`, let `f` be the unique degree `len(coords)` polynomial such that `f(x_i) = y_i` for all `i`.
+    ///
+    /// input: coords, x
+    ///
+    /// output: (f(x), Prod_i (x - x_i))
+    ///
+    /// constrains all x_i and x are distinct
+    fn lagrange_and_eval(
+        &self,
+        ctx: &mut Context<F>,
+        coords: &[(AssignedValue<F>, AssignedValue<F>)],
+        x: AssignedValue<F>,
+    ) -> (AssignedValue<F>, AssignedValue<F>) {
+        let mut z = self.sub(ctx, Existing(x), Existing(coords[0].0));
+        for coord in coords.iter().skip(1) {
+            let sub = self.sub(ctx, Existing(x), Existing(coord.0));
+            z = self.mul(ctx, Existing(z), Existing(sub));
+        }
+        let mut eval = None;
+        for i in 0..coords.len() {
+            // compute (x - x_i) * Prod_{j != i} (x_i - x_j)
+            let mut denom = self.sub(ctx, Existing(x), Existing(coords[i].0));
+            for j in 0..coords.len() {
+                if i == j {
+                    continue;
+                }
+                let sub = self.sub(ctx, Existing(coords[i].0), Existing(coords[j].0));
+                denom = self.mul(ctx, Existing(denom), Existing(sub));
+            }
+            // TODO: batch inversion
+            let is_zero = self.is_zero(ctx, &denom);
+            self.assert_is_const(ctx, &is_zero, F::zero());
+
+            // y_i / denom
+            let quot = self.div_unsafe(ctx, Existing(coords[i].1), Existing(denom));
+            eval = if let Some(eval) = eval {
+                let eval = self.add(ctx, Existing(eval), Existing(quot));
+                Some(eval)
+            } else {
+                Some(quot)
+            };
+        }
+        let out = self.mul(ctx, Existing(eval.unwrap()), Existing(z));
+        (out, z)
+    }
+}
+
 impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn strategy(&self) -> GateStrategy {
         self.strategy
@@ -349,7 +998,7 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn assign_region_in<'a>(
         &self,
         ctx: &mut Context<F>,
-        inputs: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
         gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
         phase: usize,
     ) -> Vec<AssignedValue<F>> {
@@ -436,7 +1085,7 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn assign_region_last_in<'a>(
         &self,
         ctx: &mut Context<F>,
-        inputs: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        inputs: impl IntoIterator<Item = QuantumCell<F>>,
         gate_offsets: impl IntoIterator<Item = (isize, Option<[F; 3]>)>,
         phase: usize,
     ) -> AssignedValue<F> {
@@ -525,8 +1174,8 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn inner_product<'a>(
         &self,
         ctx: &mut Context<F>,
-        a: impl IntoIterator<Item = QuantumCell<'a, F>>,
-        b: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
     ) -> AssignedValue<F> {
         // we will do special handling of the cases where one of the vectors is all constants
         match self.strategy {
@@ -541,8 +1190,8 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn inner_product_with_sums<'a>(
         &self,
         ctx: &mut Context<F>,
-        a: impl IntoIterator<Item = QuantumCell<'a, F>>,
-        b: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
     ) -> Box<dyn Iterator<Item = AssignedValue<F>>> {
         let mut b = b.into_iter().peekable();
         let flag = matches!(b.peek(), Some(&Constant(c)) if c == F::one());
@@ -559,8 +1208,8 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn inner_product_left<'a>(
         &self,
         ctx: &mut Context<F>,
-        a: impl IntoIterator<Item = QuantumCell<'a, F>>,
-        b: impl IntoIterator<Item = QuantumCell<'a, F>>,
+        a: impl IntoIterator<Item = QuantumCell<F>>,
+        b: impl IntoIterator<Item = QuantumCell<F>>,
         a_assigned: &mut Vec<AssignedValue<F>>,
     ) -> AssignedValue<F> {
         match self.strategy {
@@ -667,8 +1316,8 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn sum_products_with_coeff_and_var<'a, 'b: 'a>(
         &self,
         ctx: &mut Context<F>,
-        values: impl IntoIterator<Item = (F, QuantumCell<'a, F>, QuantumCell<'a, F>)>,
-        var: QuantumCell<'a, F>,
+        values: impl IntoIterator<Item = (F, QuantumCell<F>, QuantumCell<F>)>,
+        var: QuantumCell<F>,
     ) -> AssignedValue<F> {
         // TODO: optimize
         match self.strategy {
@@ -691,7 +1340,7 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
                             Some((va, vb))
                         } else if c != F::zero() {
                             let prod = self.mul(ctx, va, vb);
-                            Some((QuantumCell::ExistingOwned(prod), Constant(c)))
+                            Some((QuantumCell::Existing(prod), Constant(c)))
                         } else {
                             None
                         }
@@ -708,10 +1357,14 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn select(
         &self,
         ctx: &mut Context<F>,
-        a: QuantumCell<F>,
-        b: QuantumCell<F>,
-        sel: QuantumCell<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+        sel: impl Into<QuantumCell<F>>,
     ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+        let sel = sel.into();
+
         let diff_val: Value<F> = a.value().zip(b.value()).map(|(a, b)| *a - b);
         let out_val = diff_val * sel.value() + b.value();
         match self.strategy {
@@ -761,10 +1414,14 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
     fn or_and(
         &self,
         ctx: &mut Context<F>,
-        a: QuantumCell<F>,
-        b: QuantumCell<F>,
-        c: QuantumCell<F>,
+        a: impl Into<QuantumCell<F>>,
+        b: impl Into<QuantumCell<F>>,
+        c: impl Into<QuantumCell<F>>,
     ) -> AssignedValue<F> {
+        let a = a.into();
+        let b = b.into();
+        let c = c.into();
+
         let bc_val = b.value().zip(c.value()).map(|(b, c)| *b * c);
         let not_bc_val = bc_val.map(|x| F::one() - x);
         let not_a_val = a.value().map(|x| *x - F::one());
@@ -822,9 +1479,9 @@ impl<F: ScalarField> GateInstructions<F> for FlexGateConfig<F> {
                 ctx,
                 vec![
                     Constant(F::zero()),
-                    Existing(bit_cell),
-                    Existing(bit_cell),
-                    Existing(bit_cell),
+                    Existing(*bit_cell),
+                    Existing(*bit_cell),
+                    Existing(*bit_cell),
                 ],
                 vec![(0, None)],
             );
