+++ halo2-ecc/src/bigint/add_no_carry.rs
@@ -14,7 +14,7 @@ pub fn assign<F: PrimeField>(
         .limbs
         .iter()
         .zip(b.limbs.iter())
-        .map(|(a_limb, b_limb)| gate.add(ctx, Existing(a_limb), Existing(b_limb)))
+        .map(|(a_limb, b_limb)| gate.add(ctx, Existing(*a_limb), Existing(*b_limb)))
         .collect();
 
     OverflowInteger::construct(out_limbs, max(a.max_limb_bits, b.max_limb_bits) + 1)
@@ -28,7 +28,7 @@ pub fn crt<F: PrimeField>(
 ) -> CRTInteger<F> {
     assert_eq!(a.truncation.limbs.len(), b.truncation.limbs.len());
     let out_trunc = assign::<F>(gate, ctx, &a.truncation, &b.truncation);
-    let out_native = gate.add(ctx, Existing(&a.native), Existing(&b.native));
+    let out_native = gate.add(ctx, Existing(a.native), Existing(b.native));
     let out_val = a.value.as_ref().zip(b.value.as_ref()).map(|(a, b)| a + b);
     CRTInteger::construct(out_trunc, out_native, out_val)
 }
