+++ halo2-ecc/src/bigint/big_is_zero.rs
@@ -13,7 +13,7 @@ pub fn positive<F: PrimeField>(
     assert_ne!(k, 0);
     debug_assert!(a.max_limb_bits as u32 + k.ilog2() < F::CAPACITY);
 
-    let sum = gate.sum(ctx, a.limbs.iter().map(Existing));
+    let sum = gate.sum(ctx, a.limbs.iter().copied().map(Existing));
     gate.is_zero(ctx, &sum)
 }
 
@@ -30,7 +30,7 @@ pub fn assign<F: PrimeField>(
     let mut partial = gate.is_zero(ctx, a_limbs.next().unwrap());
     for a_limb in a_limbs {
         let limb_is_zero = gate.is_zero(ctx, a_limb);
-        partial = gate.and(ctx, Existing(&limb_is_zero), Existing(&partial));
+        partial = gate.and(ctx, Existing(limb_is_zero), Existing(partial));
     }
     partial
 }
@@ -42,5 +42,5 @@ pub fn crt<F: PrimeField>(
 ) -> AssignedValue<F> {
     let out_trunc = assign::<F>(gate, ctx, &a.truncation);
     let out_native = gate.is_zero(ctx, &a.native);
-    gate.and(ctx, Existing(&out_trunc), Existing(&out_native))
+    gate.and(ctx, Existing(out_trunc), Existing(out_native))
 }
