+++ halo2-ecc/src/bigint/select.rs
@@ -16,7 +16,9 @@ pub fn assign<F: PrimeField>(
         .limbs
         .iter()
         .zip(b.limbs.iter())
-        .map(|(a_limb, b_limb)| gate.select(ctx, Existing(a_limb), Existing(b_limb), Existing(sel)))
+        .map(|(a_limb, b_limb)| {
+            gate.select(ctx, Existing(*a_limb), Existing(*b_limb), Existing(*sel))
+        })
         .collect();
 
     OverflowInteger::construct(out_limbs, max(a.max_limb_bits, b.max_limb_bits))
@@ -35,7 +37,9 @@ pub fn crt<F: PrimeField>(
         .limbs
         .iter()
         .zip(b.truncation.limbs.iter())
-        .map(|(a_limb, b_limb)| gate.select(ctx, Existing(a_limb), Existing(b_limb), Existing(sel)))
+        .map(|(a_limb, b_limb)| {
+            gate.select(ctx, Existing(*a_limb), Existing(*b_limb), Existing(*sel))
+        })
         .collect();
 
     let out_trunc = OverflowInteger::construct(
@@ -43,7 +47,7 @@ pub fn crt<F: PrimeField>(
         max(a.truncation.max_limb_bits, b.truncation.max_limb_bits),
     );
 
-    let out_native = gate.select(ctx, Existing(&a.native), Existing(&b.native), Existing(sel));
+    let out_native = gate.select(ctx, Existing(a.native), Existing(b.native), Existing(*sel));
     let out_val = a.value.as_ref().zip(b.value.as_ref()).zip(sel.value()).map(|((a, b), s)| {
         if s.is_zero_vartime() {
             b.clone()
