+++ halo2-ecc/src/bn254/final_exp.rs
@@ -173,11 +173,11 @@ impl<F: PrimeField> Fp12Chip<F> {
         // compute `g0 + 1`
         g0.coeffs[0].truncation.limbs[0] = fp2_chip.range().gate.add(
             ctx,
-            Existing(&g0.coeffs[0].truncation.limbs[0]),
+            Existing(g0.coeffs[0].truncation.limbs[0]),
             Constant(F::one()),
         );
         g0.coeffs[0].native =
-            fp2_chip.range().gate.add(ctx, Existing(&g0.coeffs[0].native), Constant(F::one()));
+            fp2_chip.range().gate.add(ctx, Existing(g0.coeffs[0].native), Constant(F::one()));
         g0.coeffs[0].truncation.max_limb_bits += 1;
         g0.coeffs[0].value = g0.coeffs[0].value.as_ref().map(|v| v + 1usize);
